requirements.txt
Flask
requests
pymongo
flask-pymongo
python-nmap
